DBproject2/requirements.txt
asgiref==3.5.2
click==8.1.3
colorama==0.4.6
distlib==0.3.6
Django==4.1.3
django-cors-headers==3.13.0
djangorestframework==3.14.0
filelock==3.8.0
Flask==2.2.2
Flask-MySQLdb==1.0.1
Flask-SQLAlchemy==3.0.2
greenlet==2.0.1
gunicorn==20.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
mysqlclient==2.1.1
platformdirs==2.5.4
PyMySQL==1.0.2
pytz==2022.6
PyYAML==6.0
SQLAlchemy==1.4.44
sqlparse==0.4.3
tzdata==2022.6
virtualenv==20.16.7
Werkzeug==2.2.2
